Revert "[SECURITUM]: No invalidation of the session after logout (#869)"
This reverts commit 1e9feb13b686cc89c5bcade83eecd18045e3c5a0.
