+++ .gitignore
@@ -11,4 +11,3 @@
 /node_modules
 /test-files
 /test/temp-test
-CLAUDE.md
