+++ src/apps/main/auth/handlers.ts
@@ -6,7 +6,7 @@ import eventBus from '../event-bus';
 import { setupRootFolder } from '../virtual-root-folder/service';
 import { getWidget } from '../windows/widget';
 import { checkUserData, createTokenSchedule } from './refresh-token';
-import { canHisConfigBeRestored, encryptToken, getUser, setCredentials } from './service';
+import { canHisConfigBeRestored, encryptToken, getUser, logout, setCredentials } from './service';
 import { logger } from '@/apps/shared/logger/logger';
 import { initSyncEngine } from '../remote-sync/handlers';
 import { cleanAndStartRemoteNotifications } from '../realtime';
@@ -29,6 +29,8 @@ export function getIsLoggedIn() {
 export function onUserUnauthorized() {
   eventBus.emit('USER_WAS_UNAUTHORIZED');
   eventBus.emit('USER_LOGGED_OUT');
+
+  logout();
   Logger.info('[AUTH] User has been logged out because it was unauthorized');
   setIsLoggedIn(false);
 }
@@ -42,6 +44,8 @@ export async function checkIfUserIsLoggedIn() {
     });
     eventBus.emit('USER_LOGGED_OUT');
     setIsLoggedIn(false);
+
+    logout();
   }
 
   if (!isLoggedIn) return;
@@ -77,6 +81,8 @@ export function setupAuthIpcHandlers() {
     eventBus.emit('USER_LOGGED_OUT');
 
     setIsLoggedIn(false);
+
+    logout();
   });
 }
