+++ src/apps/main/main.ts
@@ -59,7 +59,6 @@ import { unregisterVirtualDrives } from './background-processes/sync-engine/serv
 import { setUpBackups } from './background-processes/backups/setUpBackups';
 import { setupIssueHandlers } from './background-processes/issues';
 import { setupIpcDriveServerWip } from '@/infra/drive-server-wip/out/ipc-main';
-import { AuthModule } from '@/backend/features/auth/auth.module';
 
 const gotTheLock = app.requestSingleInstanceLock();
 
@@ -168,9 +167,8 @@ eventBus.on('USER_LOGGED_OUT', async () => {
     widget.destroy();
   }
 
-  clearAntivirus();
+  await clearAntivirus();
   unregisterVirtualDrives({});
-  void AuthModule.logout();
 
   await createAuthWindow();
 });
