+++ src/apps/shared/HttpClient/client.ts
@@ -1,5 +1,6 @@
 import createClient, { Middleware } from 'openapi-fetch';
 import { paths } from './schema';
+import { logout } from '../../../apps/main/auth/service';
 import { getConfig } from '../../sync-engine/config';
 import { ipcRendererSyncEngine } from '../../sync-engine/ipcRendererSyncEngine';
 import eventBus from '@/apps/main/event-bus';
@@ -20,7 +21,7 @@ const handleOnUserUnauthorized = () => {
     ipcRendererSyncEngine.emit('USER_IS_UNAUTHORIZED');
   } else {
     eventBus.emit('USER_WAS_UNAUTHORIZED');
-    eventBus.emit('USER_LOGGED_OUT');
+    logout();
   }
 };
