--- src/backend/features/auth/auth.module.ts
@@ -1,5 +1,0 @@
-import { logout } from './services/logout.service';
-
-export const AuthModule = {
-  logout,
-};
