--- src/backend/features/auth/services/utils/reset-config.ts
@@ -1,9 +1,0 @@
-import ConfigStore from '@/apps/main/config';
-import { defaults } from '@/core/electron/store/defaults';
-import { fieldsToSave } from '@/core/electron/store/fields-to-save';
-
-export function resetConfig() {
-  for (const field of fieldsToSave) {
-    ConfigStore.set(field, defaults[field]);
-  }
-}
