--- src/backend/features/auth/services/utils/reset-credentials.ts
@@ -1,9 +1,0 @@
-import ConfigStore from '@/apps/main/config';
-import { defaults } from '@/core/electron/store/defaults';
-import { credentialFields } from '@/core/electron/store/credentials';
-
-export function resetCredentials() {
-  for (const field of credentialFields) {
-    ConfigStore.set(field, defaults[field]);
-  }
-}
