--- src/backend/features/auth/services/utils/save-config.ts
@@ -1,24 +1,0 @@
-import { getUser } from '@/apps/main/auth/service';
-import ConfigStore from '@/apps/main/config';
-import { fieldsToSave } from '@/core/electron/store/fields-to-save';
-
-export function saveConfig() {
-  const userUUID = getUser()?.uuid;
-  if (!userUUID) {
-    return;
-  }
-
-  const savedConfigs = ConfigStore.get('savedConfigs');
-
-  const configToSave: Record<string, unknown> = {};
-
-  for (const field of fieldsToSave) {
-    const value = ConfigStore.get(field);
-    configToSave[field] = value;
-  }
-
-  ConfigStore.set('savedConfigs', {
-    ...savedConfigs,
-    [userUUID]: configToSave,
-  });
-}
