--- src/core/electron/store/credentials.ts
@@ -1,3 +1,0 @@
-import { AppStore } from './app-store.interface';
-
-export const credentialFields: (keyof AppStore)[] = ['mnemonic', 'userData', 'bearerToken', 'bearerTokenEncrypted', 'newToken'] as const;
