--- src/core/electron/store/fields-to-save.ts
@@ -1,11 +1,0 @@
-// Fields to persist between user sessions
-export const fieldsToSave = [
-  'backupsEnabled',
-  'backupInterval',
-  'lastBackup',
-  'syncRoot',
-  'lastSync',
-  'deviceId',
-  'deviceUuid',
-  'backupList',
-] as const;
